fix(checker): split spread constraint helpers
