+++ crates/tsz-checker/src/query_boundaries/common.rs
@@ -1468,15 +1468,6 @@ pub(crate) fn is_array_or_tuple_type(db: &dyn TypeDatabase, type_id: TypeId) -> 
     tsz_solver::type_queries::is_array_or_tuple_type(db, type_id)
 }
 
-/// Whether `type_id`'s base constraint resolves to an array or tuple type
-/// (type-parameter / `infer` constraint chains and deferred-conditional default
-/// constraints are followed first). The caller must pass an env-evaluated type
-/// so a generic-alias `Application` like `Parameters<F>` is already expanded to
-/// its deferred-conditional body.
-pub(crate) fn base_constraint_is_array_or_tuple(db: &dyn TypeDatabase, type_id: TypeId) -> bool {
-    tsz_solver::type_queries::base_constraint_is_array_or_tuple(db, type_id)
-}
-
 pub(crate) fn is_bare_infer_placeholder(db: &dyn TypeDatabase, type_id: TypeId) -> bool {
     tsz_solver::type_queries::is_bare_infer_placeholder_db(db, type_id)
 }
